Shows the current scavenge status and scavenge history on the admin page of the UI
The admin page now shows the details of the last ten scavenges and any currently running scavenges.
These can be drilled into to view a log of the chunks that have been scavenged with the amount of space saved and time taken.
Also adds a scavengeNotificationService which polls the scavenges index stream to show a toaster message when a new scavenge starts or completes.