Allow non-admin users to log in to the admin UI
Get the user's groups on login and store them with their credentials in the ES cookie
Don't start the $scavenges poll if the user is not an admin user

Show a message in the stream browser telling non-admin users they can't view recently created streams
Don't query for streams/$all if the user is not admin
Hide 'Edit ACL' button for non-admin users on streams
Check if a stream exists and that the user has access to it before going to a searched-for stream.

Disable all tabs except dashboard, streams and queries for non-admin users
Disable nav items when they are unavailable instead of hiding them and show a tooltip explaining why it's disabled

Return to the correct state (query or projection) when going back from debugging a projection
and reload the query if you are going back to the query state from the projections debugging state.

Include the state reason for instance when a projection fails because of an authorisation exception, it's useful to know why it failed